requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
anthropic>=0.25.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
textblob>=0.18.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
newsapi-python>=0.2.7
